Extracted transfer cells to views
