+++ Wire-iOS/Sources/UserInterface/Conversation/Content/Cells/VideoMessageCell.swift
@@ -36,55 +36,25 @@ extension ZMConversationMessage {
 
 /// Displays the video message with different states
 public final class VideoMessageCell: ConversationCell {
-    private let previewImageView = UIImageView()
-    private let progressView = CircularProgressView()
-    private let playButton = IconButton()
-    private let bottomGradientView = GradientView()
-    private let timeLabel = UILabel()
-    private let loadingView = ThreeDotsLoadingView()
-    private var topMargin : NSLayoutConstraint?
+  
+    
+    private let videoMessageView = VideoMessageView()
     private let obfuscationView = ObfuscationView(icon: .videoMessage)
 
-    private let normalColor = UIColor.black.withAlphaComponent(0.4)
-    private let failureColor = UIColor.red.withAlphaComponent(0.24)
-    private var allViews : [UIView] = []
+    private var topMargin: NSLayoutConstraint!
     
     public required override init(style: UITableViewCellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
         
-        self.previewImageView.translatesAutoresizingMaskIntoConstraints = false
-        self.previewImageView.contentMode = .scaleAspectFill
-        self.previewImageView.clipsToBounds = true
-        self.previewImageView.backgroundColor = UIColor.wr_color(fromColorScheme: ColorSchemeColorPlaceholderBackground)
+        self.contentView.addSubview(self.videoMessageView)
+        self.videoMessageView.delegate = self
         
-        self.playButton.translatesAutoresizingMaskIntoConstraints = false
-        self.playButton.addTarget(self, action: #selector(VideoMessageCell.onActionButtonPressed(_:)), for: .touchUpInside)
-        self.playButton.accessibilityLabel = "VideoActionButton"
-        self.playButton.layer.masksToBounds = true
-        
-        self.progressView.translatesAutoresizingMaskIntoConstraints = false
-        self.progressView.isUserInteractionEnabled = false
-        self.progressView.accessibilityLabel = "VideoProgressView"
-        self.progressView.deterministic = true
-        
-        self.bottomGradientView.translatesAutoresizingMaskIntoConstraints = false
-        self.bottomGradientView.gradientLayer.colors = [UIColor.clear.cgColor, UIColor.black.withAlphaComponent(0.4).cgColor]
-        
-        self.timeLabel.translatesAutoresizingMaskIntoConstraints = false
-        self.timeLabel.numberOfLines = 1
-        self.timeLabel.accessibilityLabel = "VideoActionTimeLabel"
-        
-        self.loadingView.translatesAutoresizingMaskIntoConstraints = false
-        self.loadingView.isHidden = true
-
-        self.allViews = [previewImageView, playButton, bottomGradientView, progressView, timeLabel, loadingView, obfuscationView]
-        self.allViews.forEach(messageContentView.addSubview)
-        
-        CASStyler.default().styleItem(self)
+        self.contentView.addSubview(self.obfuscationView)
         
         self.createConstraints()
         var currentElements = self.accessibilityElements ?? []
-        currentElements.append(contentsOf: [previewImageView, playButton, timeLabel, progressView, likeButton, messageToolboxView])
+        currentElements.append(videoMessageView.accessibilityElements)
+        currentElements.append(contentsOf: [likeButton, messageToolboxView])
         self.accessibilityElements = currentElements
 
         setNeedsLayout()
@@ -96,38 +66,19 @@ public final class VideoMessageCell: ConversationCell {
     }
     
     open func createConstraints() {
-        constrain(self.messageContentView, self.previewImageView, self.progressView, self.playButton, self.bottomGradientView) { messageContentView, previewImageView, progressView, playButton, bottomGradientView in
+        constrain(self.messageContentView, self.videoMessageView, self.obfuscationView) { messageContentView, videoMessageView, obfuscationView in
             messageContentView.width == messageContentView.height * (4.0 / 3.0)
-            topMargin = (previewImageView.edges == messageContentView.edges).first
-            playButton.center == previewImageView.center
-            playButton.width == 56
-            playButton.height == playButton.width
-            progressView.center == playButton.center
-            progressView.width == playButton.width - 2
-            progressView.height == playButton.height - 2
-            bottomGradientView.left == messageContentView.left
-            bottomGradientView.right == messageContentView.right
-            bottomGradientView.bottom == messageContentView.bottom
-            bottomGradientView.height == 56
-        }
-        
-        constrain(bottomGradientView, timeLabel, previewImageView, loadingView) { bottomGradientView, sizeLabel, previewImageView, loadingView in
-            sizeLabel.right == bottomGradientView.right - 16
-            sizeLabel.centerY == bottomGradientView.centerY
-            loadingView.center == previewImageView.center
-        }
-
-        constrain(previewImageView, countdownContainerView, obfuscationView) { previewImageView, countDownContainer, obfuscationView in
-            countDownContainer.top == previewImageView.top + 8
-            obfuscationView.edges == previewImageView.edges
+            topMargin = (videoMessageView.edges == messageContentView.edges).first
+            obfuscationView.edges == videoMessageView.edges
         }
     }
     
     open override func update(forMessage changeInfo: MessageChangeInfo!) -> Bool {
         let needsLayout = super.update(forMessage: changeInfo)
-        
+        self.obfuscationView.isHidden = !message.isObfuscated
+
         if let fileMessageData = self.message.fileMessageData {
-            self.configureForVideoMessage(fileMessageData, initialConfiguration: false)
+            self.configureForVideoMessage(fileMessageData, isInitial: false)
         }
 
         return needsLayout
@@ -135,112 +86,36 @@ public final class VideoMessageCell: ConversationCell {
     
     override open func configure(for message: ZMConversationMessage!, layoutProperties: ConversationCellLayoutProperties!) {
         super.configure(for: message, layoutProperties: layoutProperties)
-        
+        self.obfuscationView.isHidden = !message.isObfuscated
+
         if Message.isVideoMessage(message), let fileMessageData = message.fileMessageData {
-            self.configureForVideoMessage(fileMessageData, initialConfiguration: true)
+            self.configureForVideoMessage(fileMessageData, isInitial: true)
         }
         else {
             fatalError("Wrong message type: \(type(of: message)): \(message)")
         }
     }
     
-    private func configureForVideoMessage(_ fileMessageData: ZMFileMessageData, initialConfiguration: Bool) {
-        guard let fileMessageData = message.fileMessageData else {
-            return
-        }
+    private func configureForVideoMessage(_ fileMessageData: ZMFileMessageData, isInitial: Bool) {
+        self.videoMessageView.configure(for: message, isInitial: isInitial)
         
-        message.requestImageDownload()
-        configureVisibleViews(forfileMessageData: fileMessageData, initialConfiguration: initialConfiguration)
         topMargin?.constant = layoutProperties.showSender ? 12 : 0
     }
-
-    private func configureVisibleViews(forfileMessageData fileMessageData: ZMFileMessageData, initialConfiguration: Bool) {
-        guard let state = FileMessageCellState.fromConversationMessage(message) else { return }
-        
-        var visibleViews : [UIView] = [previewImageView]
-        
-        
-        if (state == .unavailable) {
-            visibleViews = [previewImageView, loadingView]
-            self.previewImageView.image = nil
-        } else {
-            updateTimeLabel(withFileMessageData: fileMessageData)
-            
-            if let previewData = fileMessageData.previewData {
-                visibleViews.append(contentsOf: [previewImageView, bottomGradientView, timeLabel, playButton])
-                self.previewImageView.image = UIImage(data: previewData)
-                self.timeLabel.textColor = UIColor.wr_color(fromColorScheme: ColorSchemeColorTextForeground, variant: .dark)
-            } else {
-                visibleViews.append(contentsOf: [previewImageView, timeLabel, playButton])
-                self.previewImageView.image = nil
-                self.timeLabel.textColor = UIColor.wr_color(fromColorScheme: ColorSchemeColorTextForeground)
-            }
+    
+    override open var tintColor: UIColor! {
+        didSet {
+            self.videoMessageView.tintColor = self.tintColor
         }
-        
-        if state == .uploading || state == .downloading {
-            self.progressView.setProgress(fileMessageData.progress, animated: !initialConfiguration)
-            visibleViews.append(progressView)
-        }
-        
-        if let viewsState = state.viewsStateForVideo() {
-            self.playButton.setIcon(viewsState.playButtonIcon, with: .actionButton, for: .normal)
-            self.playButton.backgroundColor = viewsState.playButtonBackgroundColor
-        }
-
-        if state == .obfuscated {
-            visibleViews = [obfuscationView]
-        }
-        
-        self.updateVisibleViews(self.allViews, visibleViews: visibleViews, animated: !self.loadingView.isHidden)
     }
     
-    private func updateTimeLabel(withFileMessageData fileMessageData: ZMFileMessageData) {
-        let duration = Int(roundf(Float(fileMessageData.durationMilliseconds) / 1000.0))
-        var timeLabelText = ByteCountFormatter.string(fromByteCount: Int64(fileMessageData.size), countStyle: .binary)
-        
-        if duration != 0 {
-            let (seconds, minutes) = (duration % 60, duration / 60)
-            let time = String(format: "%d:%02d", minutes, seconds)
-            timeLabelText = time + " · " + timeLabelText
-        }
-        
-        self.timeLabel.text = timeLabelText
-        self.timeLabel.accessibilityValue = self.timeLabel.text
-    }
-    
-    override public func layoutSubviews() {
-        super.layoutSubviews()
-        self.playButton.layer.cornerRadius = self.playButton.bounds.size.width / 2.0
-    }
-    
-    // MARK: - Actions
-
-    open func onActionButtonPressed(_ sender: UIButton) {
-        guard let fileMessageData = self.message.fileMessageData else { return }
-        
-        switch(fileMessageData.transferState) {
-        case .downloading:
-            self.progressView.setProgress(0, animated: false)
-            self.delegate?.conversationCell?(self, didSelect: .cancel)
-        case .uploading:
-            if .none != fileMessageData.fileURL {
-                self.delegate?.conversationCell?(self, didSelect: .cancel)
-            }
-        case .cancelledUpload, .failedUpload:
-            self.delegate?.conversationCell?(self, didSelect: .resend)
-        case .uploaded, .downloaded, .failedDownload:
-            self.delegate?.conversationCell?(self, didSelect: .present)
-        }
-    }
-
     // MARK: - Selection
     
     open override var selectionView: UIView! {
-        return previewImageView
+        return videoMessageView
     }
     
     open override var selectionRect: CGRect {
-        return previewImageView.bounds
+        return videoMessageView.bounds
     }
     
     // MARK: - Menu
@@ -314,4 +189,11 @@ public final class VideoMessageCell: ConversationCell {
         }
     }
 
+}
+
+
+extension VideoMessageCell: TransferViewDelegate {
+    public func transferView(_ view: TransferView, didSelect action: MessageAction) {
+        self.delegate.conversationCell?(self, didSelect: action)
+    }
 }
